+++ hocon/src/test/java/de/cubeisland/engine/reflect/hocon/HoconReflectedTest.java
@@ -23,12 +23,17 @@
 package de.cubeisland.engine.reflect.hocon;
 
 import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileOutputStream;
+import de.cubeisland.engine.reflect.ReflectedTest;
+import de.cubeisland.engine.reflect.ReflectedTest2;
 import de.cubeisland.engine.reflect.Reflector;
+import de.cubeisland.engine.reflect.codec.hocon.HoconCodec;
 import de.cubeisland.engine.reflect.exception.DuplicatedPathException;
 import org.junit.Before;
 import org.junit.Test;
 
-import static de.cubeisland.engine.reflect.hocon.ReflectedFieldShadowing.ReflectedFieldShadowing2;
+import static de.cubeisland.engine.reflect.ReflectedFieldShadowing.ReflectedFieldShadowing2;
 import static org.junit.Assert.assertEquals;
 
 public class HoconReflectedTest
@@ -38,6 +43,7 @@ public class HoconReflectedTest
     private File file;
 
     private Reflector factory;
+    private HoconCodec codec;
 
     @Before
     public void setUp() throws Exception
@@ -46,25 +52,27 @@ public class HoconReflectedTest
         this.factory = new Reflector();
         this.test1 = factory.create(ReflectedTest.class);
         this.test2 = factory.create(ReflectedTest2.class);
+        codec = factory.getCodecManager().getCodec(HoconCodec.class);
     }
 
     @Test
     public void test() throws Exception
     {
-        ReflectedTest loadConfig = factory.load(ReflectedTest.class, file);
+        codec.saveReflected(test1, new FileOutputStream(file));
+        final ReflectedTest reflected = factory.create(ReflectedTest.class);
+        codec.loadReflected(reflected, new FileInputStream(file));
         file.delete();
-        assertEquals(test1.getCodec().convertReflected(test1).asString(), test1.getCodec().convertReflected(
-            loadConfig).asString());
+        assertEquals(codec.convertReflected(test1).asString(), codec.convertReflected(reflected).asString());
     }
 
     @Test
     public void test2() throws Exception
     {
-        test2.save(file);
-        ReflectedTest2 loadConfig = factory.load(ReflectedTest2.class, file);
+        codec.saveReflected(test2, new FileOutputStream(file));
+        final ReflectedTest2 reflected = factory.create(ReflectedTest2.class);
+        codec.loadReflected(reflected, new FileInputStream(file));
         file.delete();
-        assertEquals(test2.getCodec().convertReflected(test2).asString(), test2.getCodec().convertReflected(
-            loadConfig).asString());
+        assertEquals(codec.convertReflected(test2).asString(), codec.convertReflected(reflected).asString());
     }
 
     @Test(expected = DuplicatedPathException.class)
