--- hocon/src/test/java/de/cubeisland/engine/reflect/hocon/ReflectedTest2.java
@@ -1,43 +1,0 @@
-/**
- * The MIT License
- * Copyright (c) 2013 Cube Island
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-package de.cubeisland.engine.reflect.hocon;
-
-import de.cubeisland.engine.reflect.Section;
-import de.cubeisland.engine.reflect.codec.hocon.ReflectedHocon;
-
-public class ReflectedTest2 extends ReflectedHocon
-{
-    public Section3 section3 = new Section3();
-
-    public static class Section3 implements Section
-    {
-        public String section = "actually not a section";
-    }
-
-    public Section2 section = new Section2();
-
-    public static class Section2 implements Section
-    {
-        public String aValue = "aString";
-    }
-}
